requirements.txt
Flask==3.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9

pytest==8.2.0
python-dotenv==1.0.1

isort==5.13.2
flake8==7.0.0
black==24.4.2
